Refactor Ask screen by collapsing voice/render mixins
